(SIO-2190) Replace NoArgsCommand with BaseCommand

+++ oioioi/ipauthsync/management/commands/ipauthsync_clean.py
@@ -1,14 +1,17 @@
-from django.core.management.base import NoArgsCommand
+from django.core.management.base import BaseCommand, CommandError
 from django.db import transaction
 
 from oioioi.ipauthsync.models import IpAuthSyncedUser
 from oioioi.ipdnsauth.models import IpToUser
 
 
-class Command(NoArgsCommand):
+class Command(BaseCommand):
     help = "Removes synced entries from the IP database."
 
-    def handle_noargs(self, **options):
+    def handle(self, *args, **options):
+        if args:
+            raise CommandError("This command doesn't accept any arguments!")
+
         with transaction.atomic():
             IpToUser.objects.filter(ipauthsynceduser__isnull=False) \
                     .delete()
